// src/models/documentation.ts
import { CommandOptionConfig } from "./options";

export interface CommandDocumentation {
  readonly name: string;
  // FIXME: should be autogenerated from "raw options"
  readonly usage: string;
  readonly description: string;
  readonly details: string;
  readonly options: Map<string, CommandOptionConfig>;
}